会员积分针对代码格式、功能优化 1.修改积分抵扣单位

+++ yudao-module-member/yudao-module-member-biz/src/main/java/cn/iocoder/yudao/module/member/dal/dataobject/point/MemberPointConfigDO.java
@@ -37,7 +37,7 @@ public class MemberPointConfigDO extends BaseDO {
      *
      * 1 积分抵扣多少分
      */
-    private BigDecimal tradeDeductUnitPrice;
+    private Integer tradeDeductUnitPrice;
     /**
      * 积分抵扣最大值
      */
